Fixes: regression caused by removing mouse press delay from mscene.
RevBy: Michael Hasselmann
Details:
The focus handling in panning usecases are problematic because of
the fact that the focus for text edits is set using mouse press
events instead of mouse release. This patch uses a new flag added
to Qt which allows to handle focus manually.

We are going to set focus on selected widgets only if there was no
pan gesture recognized and accepted.