Fixes: NB#214268 - Gallery, keep in (no items - USB mode) after Unplug the USB and gets crashes in Arabic language
RevBy: Mike Fabian
Details: worked on this together with Stefan Hundhammer.
the problem at hand is that we do try to send a
LanguageChange event, although the program is already
closing itself. to fix this problem, we will not
delete the static QTranslator pointers, but we will
give the qApp parent pointers to them. then when
the QApplication desctuctor is called, it deletes its
children, and also the QTranslators. but because it
is already self->closingDown(), it will not send out
QEvent::LanguageChange events in QCoreApplication::
removeTranslator() any more.