Changes: replace on disk CSS list datastructure with tree
RevBy: Peter Penz, Stanislav Ionascu

Details:
So far we had to iterate over all selectors to check if they are matching
an object we want to style. Now we have a tree datastructure which only
returns selectors matching the style class name and the object/style
name. The datastructure is created inside to be mapped cache files and
not created at runtime.
This is only the first step for further optimizations. The whole matching
in MStyleSheet can/should be moved into the tree.
This commit also prepares for further optimizations in style creation and
orientation change performance.