Changes: speed up doxygen doc generation 30 times.
RevBy: Adrian Yanes
Details: the FILE_VERSION_FILTER statement is executed
for every file for which documentation should be
generated. This patch chenges it from an expensive
shell script to an echo call.

For me this speeds up doc generation from 9:59 minutes
to 20 seconds. Without a version string at all the
doc creation would take 14 to 16 seconds. So i think
this overhead is acceptable.