Fixes: NB#220018 - chainTask servicefw property does not work with apps running through applauncherd
RevBy: Stefan Hundhammer

Details: The problem here was that the MComponentData::
chainDataStack was not containing any data when the
MApplicationWindow (MAW) was created. and the MAW did only
look for the chainData in its constructor. So a MAW was
created by the booster, and it was cached. and the
chainData that was received from dbus later did not
have a chance to get through to the cached MAW.

To fix that i added a signal to the MComponentData
that gets emitted when the chainData changes, and i
do listen to that event in the MAW when no chainData
was found in the first place. Also i do disconnect
from that signal when the chainData changed, to make
sure we do not get multiple MAW instances that all
want to handle that chainData change.