Changes: fix ut_mcalendar to work both with libicu 4.4 and 4.6
RevBy: Miroslav Safr, Holger Schröder

Details:

libicu 4.6 now has data for Finnish locale not only
for gegorian calendar but also for the calendar types
"buddhist", "islamic", "japanese", and "roc".

Because we override the separator between hours and minutes to be “:”
instead of “.” for Finnish locale, we need to do it consistently for
all calendar types which have such patterns defined in libicu.

See also:

commit 26a9dd6196c950cc7325643ffb95f97d45db9729
Author: Mike FABIAN <mike.fabian@basyskom.de>
Date:   Tue Sep 21 16:16:18 2010 +0200

    Fixes: NB#193497, Clock; Unusual time display separator (.) in Finnish.

    RevBy: Berthold Krevert

    Details:

    Use colon as the separator for time for Finnish.