Changes: cache style sheets for style classes
RevBy: Peter Penz, Stanislav Ionascu

Details:
For every created widget we were traversing the matching style hierarchy
to collect the matching stylesheets. When creating a lot of widgets this
is a time consuming operation. Now when a specific style is created
for the first time the matching stylesheets are saved.