Changes: reduce allocations in MStyleSheetPrivate
RevBy: Peter Penz, Stanislav Ionascu

Details:
So far we were creating an MOriginContainer for every attribute of
a style. When building the actual style we were creating even more
style containers.
A different approach is used now to fill the style: All selectors
are sorted by priority. For this we need to allocate one struct
per selector. When we fount all matching selectors we iterate over
all properties and fill them with the highest selector having a given
attribute.