Changes: add xml path to m-servicefwgen parameters
RevBy: Armin Berres
Details: Before this patch m-servicefwgen did always
create the files it generates in the directory where
the xml file was. Because of that you had to copy
xml files that were installed in the system to your
build directory before using m-servicefwgen.

Also some projects were just generating the adaptor
files with qdbusxml2cpp instead of m-servicefwgen.

This patch allows the user to give the full path
of the xml interface file to use. That way the
interface files installed in the system can be used
to generate the needed files.