Changes: Reset IM area on orientation change because the correct value is not yet updated.
RevBy: Daniel d'Andrada
Details:
Application orientation and software input panel (SIP) area are out of sync at
the time of orientation change. It takes time for the new SIP area to arrive
(via d-bus). This change resets the SIP area to null rect in MInputWidgetRelocator
when waiting for the updated area to arrive. During this time, however, further
relocations are not postponed since the new area may never come. Relocations are
just done as if there were no SIP on the screen at the moment.

This is sort of a workaround. A proper fix would require API changes
and we are probably good with this. See comments in code.
- Store orientation angle on rotation begin instead of finished, in MInputWidgetRelocator.