Changes: lazily delete styles
RevBy: Peter Penz, Stanislav Ionascu

Details:
While creating widgets we often run into a situation where a style
is only created for a short while. This might happen quite often
for exactly the same style if a widget is created in the same way
mutliple times. As the reference cound always dropped to zero the
style was immediately freed so far and had to be recreated.
Now we refrain from deleting the styles until either the application
is minimized or terminated.
In the end creating the styles which are destroyed immediately
afterwards should be avoided but this is not always easily possible.

The LeakedStyle calls is removed and merged with StyleCache.