Fixes: NB#229937 - Slider knob still slightly truncated at the min/max positions
RevBy: Eduardo Lima Mitev

Details:
MSliderGroove::updateHandlePos() uses only the current value
to select the handle's position. MSliderGroove::screenPointToValue()
is adjusted by half of the handle's width, not the groove margin;
this is so because the position of the handle's centre is what
determines the current value. MSliderGroove::paint() places the
groove progress indicator in the center of the handle for the same
reason.