Changes: fix mmoc handling of qt moc exit codes
RevBy: Stefan Hundhammer
Details: Before mmoc was silently ignoring the
error code that was returned by qt moc, thereby
hiding moc problems in the build. These errors
only were seen when the moc file to compile was
not existing, because qt moc refused to create
it because the syntax in the .h file was broken.