Fixes: NB#232757 - Arabic Gallery, Incorrect location of the percentage symbol"%" in the string §qtn_gall_resize_button_value_2§.
RevBy: John Tapsell, Holger Schröder

Details:

For formatted numbers, percent values, and currency values, add U+200F
RIGHT-TO-LEFT MARK before the first character of type EN or AN and
behind the last character of type EN or AN in the formatted result.

This makes the number display the non-numeric characters like the
minus character, the percent sign, and the currency symbols display
automatically on the correct side of the number, even if the formatted
number is displayed on its own without any Arabic text neither before
nor after it.