Changes: MWindow - Don't try to switch graphicssystems if one was specified
RevBy: Armin Berres
Details:
When the meego graphics system is present, we can and should only
switch between 'meego' and 'raster' if we are running the 'runtime'
graphics system. i.e., if no graphics system was specified in any way.
Be it via -graphicssystem command line parameter, via environment variable
or via QApplication::setGraphicsSystem().

This means that when an application explicitly says that it wants the
meego graphics system (hardware accelerated back end) we won't try to
switch it to sofware rendering (raster) when a window gets minimized.

Enough for "myapp -graphicssystem meego", but it will still crash
for "myapp -graphicssystem raster". To fix the latter changes elsewhere
are necessary since it shouldn't consider itself to be running in
"hardware rendering mode", as it currently does.