Changes: replace "const char*" with QByteArray in hash
RevBy: Stanislav Ionascu

Details:
The assumption when writing this code was that the same string
always is represented by the same pointer as the pointer is
coming from the respective meta object. This assumption does not
hold anymore if we pass in arbitrary strings used e.g. to do style
preloading.