Changes: Removing ft_mwindow. Tests removed code.
RevBy: TrustMe
Details: This test was checking if libmeegotouch wasn't
wrongly removing _NET_WM_STATE_SKIP_TASKBAR since libmeegotouch
used to set and remove this property. Since commit
85250c0a510b4f830f563e7e9cd4841e047eaa9 libmeegotouch no longer uses
this property at all. Therefore this functional test is no longer needed.

Besides that this test sleeps for a total of 3 seconds, slowing down significantly
the time it takes to run the test suit. Tests should be *fast* to run.

It also relies on the presence of a window manager and will simply fail
is one is not running (which is a common situation when developing in scratchbox).