Changes: register dbus service as soon as possible
RevBy: Peter Penz

Details:
If an application is already running the dbus service brings it
to the front and terminates the new instance. This should happen
as fast as possible. By moving the service creation before connecting
to the feedback daemon and creating the theme the time needed to
bring up a running application is reduced.