Changes: communicate most used pixmaps even when not yet completely loaded
RevBy: Peter Penz, Adrián Yanes, Miguel Gomez

Details:
So far until the most used pixmaps have not been loaded completely they
have not been communicated to the started apps. Now when an application
starts the so far loaded pixmaps are immediately communicated.
Additionally pixmaps are immediately removed from the list of to be
loaded most used pixmaps when a client requests the pixmap. This reduces
the time until all pixmaps are loaded.