Changes: Added Stub and StubMap classes for unit testing purposes
RevBy: Daniel d'Andrada
Details:
Store instances of stubs in the StubMap so that they can
each be set their individual return values for their stubbed
methods. Proxies should register themselves into the StubMap in
their constructor with an unique key, that is used on the
unit test side to get hold of them for return value setup.