Changes: optimize themedaemon protocol
RevBy: Peter Penz

Details:
Packages will be first streamed into a QByteArray and copied to the
network socket in one go. When reading the whole packet data will
be read from the stream and deserialized afterwords.
This way we do not need to constantly check if the data we want to
read is already available and can directly use the Qt streaming
operators.
For huge packages we see a speedup of ~50%.