Changes: MSheet - system sheet rotates on its own accord
RevBy: Jarosław Jaryszew
Details: MSheet's stand alone window is actually considered
to be the "current application window" by mcompositor. Thus setting this
"followsCurrentApplicationWindowOrientation" property means that system sheet
would be following its own orientation in a circular dependency.

Since it's also fully opaque, you cannot see the window behind it at all.
Thus it's not important for the system sheet to follow the orientation
of the preceding window. Maybe it could just have the orientation of
that preceding window when appearing, but that can come up in a
separate commit.