Changes: MSceneManager - Scene windows behind sheet get hidden
RevBy: Marcin Miklas
MR: 1664

Details:
A sheet by definition is fully opaque and takes the entire screen.
Therefore nothing behind it can be seen by a GraphicsView. But in
spite of that all those iems behind the sheet still get drawn.

As an optimization we hide all those items. A bonus is that they
will get a more accurate information on their visibility and will
be able to act accordingly by stopping animations, etc.