Changes: fullscreen transition animation for MApplicationPage and MSheet.
RevBy: Daniel d'Andrada
Details: MApplicaitonPage and MSheet geometry is now animated during statusbar
hiding/showing animation. MApplicationPage exposedContentRect is updated on
disappeared/appeared signal from statusBar and takes into accout page->geometry().

MApplicationPage therefore is no longer always fullscreen as it used to be.
Now, instead of having the status bar covering the top part of the page we have
the page being resized and repositioned to fit in the remaining aread below
the status bar. Therefore we are assuming that the status bar is *always* fully
opaque. If in the future we happen to have cases where the status bar is translucent
then we will have to add logic to keep the page fullscreen in such situations.