Fixes: NB#239110 - ft_theme::testViews segfaults
RevBy: Stanislav Ionascu, Peter Penz
MergeRequest: 1682

Details:
When a library is unloaded and its style creators are unregistered we
must remove all styles created by the style creators from the cache.
Otherwise deleting the styles later will lead to a crash as the code
for the d'tor is not available anymore.