Revert "Fixes: NB#240696 - [TASK] Theme should be able to retrieve application icon from default application icon folder"
This reverts commit 1f0deec1531db91ffaaf2c5c2cad4e34dc9ebc7f.

Conflicts:
	src/corelib/theme/mthemedaemon.cpp

RevBy: Armin Berres, Daniel d'Andrada.

Details: The commit is reverted as the system wide icons folder, is not part of meegotouch themes, thus it makes no sense to have them by default in the search paths of the themedaemon.
In case if applications want to access the resources in those folders (or any other app specific folders), meegotouch provides MTheme::addPixmapDirectory API.