Fixes: NB#228944 - Sharing, Image size dialog in Publish to service UI keeps displays Latin Number instead of Hindi number in Arabic language
RevBy: John Tapsell

Details:

The problem occurs because one cannot chose which digit system Qt uses
for localized digits in QLocale. Therefore, to get the localized
digits which may or may not be used in expansions of placeholders
like %Ln, %L1, ... we use the workaround of changing the locale
which is actually passed to QLocale::setDefault() to a locale
which uses number formatting which comes close to what we want.