Changes: MStatusAreaSlideAnimation - used clippedPaintOffset instead of paintOffset.
RevBy: TrustMe
Details: Using paintOffset will move the status bar downwards. For the case
where this animation is done in a scene that has a page it's not a problem since
the status bar will slip under it. Thus the ilusion will work.
But the main user of this animation will be the translucent top-level window that
systemui shows on top the the foreground application window to display the banner
of a notification. If we didn't clip the status bar movement we would see the status
bar moving downwards above the page of that application window since they are in separate
QGraphicsScenes.