Changes: MLabel - make the minimum size equal to the preferred size if we have no word wrap or eliding enabled
RevBy: Peter Penz
Details: Fixes NB#236032 Subject label NOT shown completely in Header
info view if length exceeds   among others.

The idea is that if we have two labels side by side, one with eliding
and one without, we want the one without eliding to have all the space,
and the one with eliding to be the one that shrinks.