Fixes: NB#248153 - Flick cancelling behavior is off
RevBy: Daniel d'Andrada
Details: The swipe recognizer is recording the distance that was
covered by the gesture and in case the previously recorded distance
was larger than the new one, the gesture is cancelled. This covers
the situation when the user moves his finger in one direction and
at some point he starts moving it back.

The rootcause of this bug was that the recognizer stored previous
distance without checking if the user already started moving his finger
back or not.