Fixes: NB#248584 - Disable mirroring in Arabic build causes incomplete mirroring effect in bidirectional text.
RevBy: Stefan Hundhammer

Details:

MLabelViewPrivate::autoSetTextDirection() still needs to set the
text direction correctly, even if mirroring for Arabic is disabled.
Only the horizontal alignment should change if mirroring for Arabic
is disabled. The text direction must be set correctly, if not
the results of the bidi reordering are wrong.