Changes: Optimize emission of MOrientationChangeEvent
RevBy: Marcin Miklas
Details:
Currently when the orientation changes MScenenManager sends a
MOrientationChangeEvent to every single QGraphicsWidget in the scene.

In a regular drill-down scenario that will include all the previous pages that
you can reach by pressing the back button.

The optimization idea is to send MOrientationChangeEvent events only to
widgets in scene windows whose state is different than
MSceneWindow::Disappeared. And when a scene window is about to leave its
Disappeared state MSceneManager checks if it has an outdated orientation and if
so sends MOrientationChangeEvent events for its widgets to get them up to date.

When a widget gets an MOrientationChangeEvent it updates and re-apply its
style.

This optimization will reduce the total time it takes for an orientation change
to happen. More specifically, it will reduce its setup time, i.e. the time it
takes between MOrientationTracker tells the new orientation and the orientation
animation starting to run.

Fixes: NB#249745 -[TASK] Optimize emission of MOrientationChangeEvent