Fixes: NB#241217 - Startup wizard language not in SIM language
RevBy: Stefan Hundhammer
MR: 1761

Details:
for some reason the notify_trampoline callback of the gconf
implementation is not called in our case. Perhaps someone
has time and energy to find this problem in GConf. Until
then lets have this workaround. In addition to setting a
value in GConf we also send local value updates to the other
MGConfItem instances that are used in this application.