Changes: fix crash in sbox in MGConfItem dtor
RevBy: Stefan Hundhammer

Details: The problem was that the static QHash
was deleted before the static MLocale was deleted
by the MStaticLocaleDestroyer. As the order in which
static objects are deleted is not defined, i had
to change this from a static QHash to a static
pointer to a QHash.