Changes: cleaning & ordering MBanner code. RevBy: Felipe Erias Morandeira, Armin Aberres MR: 1779 Details:
Due to deprecate code for MInfoBanner we kept some old name convetions
to handle the new MBanner.

The animations in MBanner are handled by the stylename. At the same
a new namespace is introduced to have a clean way to specify the stylename
of the MBanner through the namespace "MBannerType".