Fixes: NB#231817 - Image improperly aligned in Sheet from photo widget (given scenario)
RevBy: Marcin Miklas, Stanislav Ionascu
MergeRequest: 1786

ModifiedBy: Stanislav Ionascu
Modified: Preserve the fullscreen transition animation, relocate the photo in the full screen sheet, when the width or height, of the full screen photo sheets central widget, is changed.
The transition animation is destroyed with sheets photo, and QPointer is used to track it.

Details: The photo was not following full screen sheets geometry.