Fixes: NB#242154 - Tracker does not decode language encoding scheme
RevBy: Stefan Hundhammer

Details:

If “zh” is set in setDeclaredLocale(), i.e. the variant of Chinese
is unknown, simplified Chinese seems a bit more likely. On top of
that, the settings application sets only “zh” for simplified Chinese
and the translations for simplified Chinese are also in files like
“foo_zh.qm” which makes simplified Chinese more likely when only
“zh” is set on the device.

This change increases the confidence level of GB18030 to 39 and of
Big5 to 38 if only setDeclaredLocale("zh") is set and both Big5
are possible for the input data.