Changes: improve Ut_MApplicationPage::testRememberPosition()
RevBy: TrustMe
Details: It was relying on the fact that the view of m_subject
was already created. The view assigns a widget to the viewport
upon creation. If that doesnt happen the call to
pannableViewport->setPosition() will cause a crash.

It was also doing a lousy simulation of page appearing and disappearing
by just emitting the related signals. Now it properly appears and
disappears.