Fixes: NB#248658 - MPhysics2dPanning: Panning is restarted when a very still finger is lifted
RevBy: Daniel d'Andrada
Details: The calculation of speed after the movement has stopped does not
take into account a situation when the user holds his finger very still
for a moment before releasing it. The viewport velocity after that moment
should be set to zero but there were no movement events that could be used
to find out that the speed dropped to zero.

This patch adds a QElapsedTimer which is used to recognize such situations.