Fixes: NB#243647 - Text inside MMS bubble containing long text is cropped from top
RevBy: Stanislav Ionascu, Joaquim Rocha
MergeRequest: 1747

ModifiedBy: Stanislav Ionascu
Modifications: Removed dependency on scene window in the parent hierarchy.
The tile height does not depend on the current orientation, but rather on the max between resolution height and width.
Removed few orientation dependent checks.

Details: The coordinate-handling of the scene has been fixed to work for all rotation-angles (fixed in 96623160ce3e04d66237746585c1da9f7b894e21).
It is not mandatary anymore that the first top-tile starts at y-position 0 to show it, also only a partly visible top-tile will be shown.
The calls to MLabel::update() have been removed as they are unnecessary overhead: MLabelViewRich::updateTilesPosition() is only invoked in the context of the drawing operation and does not lead to an outdated MLabel content.

MLabelView got private methods accessible for the unit-tests to be able to check for regressions.