Changes: Updating list of year numbers in copyright notices
RevBy: TrustMe
Details:
Quoting http://www.gnu.org/prep/maintain/html_node/Copyright-Notices.html
"""
When you add the new year, it is not required to keep track of which files have
seen significant changes in the new year and which have not. It is recommended
and simpler to add the new year to all files in the package, and be done with
it for the rest of the year.
"""