New: font preloading support
RevBy: Stanislav Ionascu

Details:
The invoker is suppposed to initialize as many things as possible
which are shared between several applications. To avoid cluttering
the invoker code with this logic a new library "meegotouchpreloader"
is added. this library is dlopened by the invoker and does the needed
initializations for meego touch applications.
In the first step the library fills the internal font cache of Qt to
get rid of delays when a speficic font is requested for the first
time.
Later on this library is also supposed to precreate MStyle objects.