Changes: Allow to draw diacritics inside the margins for rich-texts
RevBy: Stanislav Ionascu
MergeRequest: 1819

ModifiedBy: Stanislav Ionascu
Modified: Changed geometry().y() call to pos().y() for better readability.

Details: For simple-texts the margins of a label may already be used for drawing diacritics (this is useful for fonts where the font-ascent is too small to contain diacritics). The patch assures that the same behavior is given also for rich-texts.

The unit-tests have been improved and exposed an issue in the vertical alignment of the tiles that has been fixed too.