Changes: Ensure that item focused via code gains focus once his MWindow is activated Fixes: NB#246452
RevBy: Jarno Malmari
Details:
Ensure that if application code calls item->setFocus() that item
will indeed gain focus once the MWindow that holds his scene gets activated.

The focus of QGraphicsItems inside a QGraphicsScene is a bit
disconnected from the focus of QApplication/QWidget but the former
does depend on the latter in order to work.