Restore "Changes: Optimize emission of MOrientationChangeEvent"
This reverts commit 54a76f5e8d1d66a7ce37b956a435e5d9ae206444.

RevBy: Marcin Micklas
Details:
Putting it back along with the proper fix for the problem that motivated
the withdrawal of the original commit (NB#252583).

The fix:
MSceneWindowPrivate::orientationFromLastEvent was not being properly
initialized. The case where MSceneWindow never got an MOrientationChangeEvent
before wasn't really being taken into consideration

Original details:
Currently when the orientation changes MScenenManager sends a
MOrientationChangeEvent to every single QGraphicsWidget in the scene.

In a regular drill-down scenario that will include all the previous pages that
you can reach by pressing the back button.

The optimization idea is to send MOrientationChangeEvent events only to
widgets in scene windows whose state is different than
MSceneWindow::Disappeared. And when a scene window is about to leave its
Disappeared state MSceneManager checks if it has an outdated orientation and if
so sends MOrientationChangeEvent events for its widgets to get them up to date.

When a widget gets an MOrientationChangeEvent it updates and re-apply its
style.

This optimization will reduce the total time it takes for an orientation change
to happen. More specifically, it will reduce its setup time, i.e. the time it
takes between MOrientationTracker tells the new orientation and the orientation
animation starting to run.

Fixes: NB#249745 -[TASK] Optimize emission of MOrientationChangeEvent