Fixes: NB#253948
RevBy: Dmitry Abramov

Details:
The color combo box was deleting its popup which was retrieved using a
method that instantiates it if it is null and this was causing a segfault.
This patch deletes the popup using its pointer instead of a method which,
besides preventing the crash, also makes it faster for not instantiating
an object that is going to be deleted right away.