Fixes: NB#239785
RevBy: Adrian Yanes

Details: Commit d6cccd3ec78745cbd0ba4869c80b086b2f34017c assured that the
diacritics can be drawn on the margin-area of the tiles.
But depending on the available height of the label the clipping
might still result in invisible diacritics when drawing the tiles.

The patch fixes the wrong clipping rectangle and adjusts the unit-test ut_mlabel to detect this kind of issue.