Fixes: NB#256530 - MToolbar disabled tab style is not updated after changing device orientation
RevBy: Daniel d'Andrada
Details:
There has been that hack to refresh disabled tool bar items (on rotation) while
menu is open. This enables it also when menu is not open so now all disabled
tool bar items are refreshed when orientation changes. Needed mostly with
the tabs that are very different looking in landscape and portrait.