Fixes: NB#255064
RevBy: Daniel d'Andrada
Details: Because MSceneManagerPrivate::canHaveAnimatedTransitions returns true
during application startup even when app is not on display then (but onDisplaySet == false),
we cannot let sceneManager to decide if statusBar appearance/disappearance should be animated
or not. So MApplicationWindow is updated to decide if statusBar should be animated based on isOnDisplay property.