Fixes: NB#257891 - [TASK] Decrease status bar tappable area size for sheet views
RevBy: Stanislav Ionascu
Details:
If a sheet is the top most scene window we place the status bar right below it.

The default MStatusBar style usually have a reactive margin on its bottom to
make it easier to tap on it to bring the status menu. But while a sheet is
being displayed this extra margin might partially cover the buttons on the
sheet header (or at least be very close to them), causing the user to
accidentally open the status menu when he wanted to press some sheet header
button.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/1904