Fixes: NB#258052 - MContainer and instantaneous bad position of progress indicator after using toggleProgressIndicatorVisible()
RevBy: Holger Schröder
MR: 1911

Details:
When showing the spinner in a container header, it would first show up
on the left and quickly repositioned to the right. This happened due
to the fact that it would be drawn before the layout finished adding
it. Showing it from timer with a single shot dealays the drawing a
bit and fixes this situation.