Fixes: app keeps context properties subscritions if MWindow::show() is not called during its initialization
RevBy: Daniel d'Andrada

ModifiedBy: Daniel d'Andrada
  - Rewrote the modifications done to MOrientationTracker::orientationAngle()
  - Added changes to Ut_MOrientationTracker::init() cleanup() and
    testOrientationPolicyDeviceProfileConstraints()

Details:
MOrientationTracker subscribes to sensor properties on its constructor.
When a window gets shown a check is done whether the subscription is
needed and an unsubscription will be done if not. But this check does not happen
when the window is not shown right away and results in unnecessary wakeups of
applications in background.

The patch fixes this by not subscribing to the sensor properties in the
MOrientationTracker constructor. For the MOrientationTracker::orientationAngle()
only a temporary subscription is done (if not subscribed already) in order to
return a correct value.

Two existing unit-tests had to be adjusted to execute the tests with a visible window,
as with the modified behavior no tracking is done in those cases on purpose.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/1929
Partial fix for NB#258447