Fixes: NB#247870 - MSheet -> hiding a header won't cause the centralWidget to take the headers space
RevBy: Daniel d'Andrada, Marcin Miklas
Details: Replaced QAnchorLayout by MLayout with MLinearLayoutPolicy. Items (HeaderSlot and CentralSlot) was set on two layers:
header slot layer and central slot layer. The first one was set as free item with controller as parent (and geometry from style). The second one is,
in fact, a linear layout with a spacer on top (MSheetSpacer) and the central slot on bottom. The spacer have the same weight as header slot, and its
geometry will be animating when header will be hiding.