Fixes: NB#258447 - All applications wake up in the background when changing the orientation of the device
RevBy: Daniel d'Andrada, Armin Berres

Details: The constructor of ContextProperty already calls ContextProperty::subscribe() but MContextProperty::m_isSubscribed is set to false in this case. This means that MContextProperty::isSubscribed() might return false although the property is subscribed.

The patch assures that the ContextProperty used inside MContextProperty stays unsubscribed to be consistent with m_isSubscribed. Also an unsubscribing has been added for the other ContextProperty instances that are not wrapped inside MContextProperty. This has not been required yet as each application asked for MOrientationTracker::orientationAngle() on startup where a temporary subscription/unsubscription is done, however unsubscribing is not expensive and lets better stay on the save side.