Fixes: NB#216388 - With the object menu on the foreground,able to enter characters through vkb. Fixes: NB#260245 - MObjectMenu and MDialog does not grab the focus from MTextEdit in some particular cases.
RevBy: Daniel d'Andrada
Details: QEvent::WindowBlocked, QEvent::WindowUnblocked are sent not only to currentPage but also
last blockerWindow if it exists. QEvent:WindowBlocked is sent when blocker sceneWindow is appearing
(and in case of non animated appearance when it reaches appeared state).

It covers cases such as a dialog coming on top of a sheet and that sheet contains a focused MTextEdit.
Sheet should be blocked so that its MTextEdit loses focus.