Changes: More docs for the declarative settings language
RevBy: Eetu Lehmusvuo
Details:
libmeegotouch contains only an autogenerated documentation for the declarative
settings language, which can be hard to read for someone who does not even have
an idea about what that is.

This patch gives some explanation about how to use the language parser to build
widgets from the xml files and about the format of the xml.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/1960