Fixes: NB#262925 - When background applications' visibility changes from Recents view to hidden, many applications do heavy work in the background
RevBy: Daniel d'Andrada

Details: An unnecessary update() call is done in the MPositionIndicator if the
display has been exited. This results in an unnecessary painting when e.g.
swiping from the recents view to the launcher view. In this case the
painting is very expensive as it is done with the raster-engine.

Verified with widgetsgallery

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/1964