Changes: add methods to set collator strength and use them for the index buckets
RevBy: John Tapsell, Holger Schröder
MR: 1978

Details:

needed for NB#261636

Until now the MCollator always had Quaternary strength.  This is OK
for sorting, but not always correct when determining the index
buckets. For example, a “Á” may compare less then “A” in
some locales (for example in Chinese pinyin sorting) but nevertheless
it should go into the “A” bucket because it only has a secondary
difference to “A”.

Similar, the Japanese characters “ゕ” and “ヵ” compare smaller
then “か” but have only a secondary difference to “か” and should
thus go into the “か” bucket.

Using a collator with only Primary strength to find the correct
buckets fixes this problem.