Changes: Fix failing tests for ut_mlabel and ut_allwidgets
RevBy: Armin Berres

Details: To bypass the issues with the Nokia-font MLabel allows (in opposite to other widgets) to draw inside the margin-area. Recently some margin-values have been changed which break the unit-tests ut_mlabel and ut_allwidgets. The patch assures that the MLabel-related-tests run with no margins.

As soon as the Nokia-font got fixed and the workaround in MLabel has been removed those CSS-changes for MLabelStyle can be removed too for the unit-tests.