Changes: Fix unit-tests ut_allwidgets and ut_mlabel for the scratchbox
RevBy: Armin Berres

Details: Because of bug 253936 the unit-tests ut_allwidgets and ut_mlabel fail when running them in the scratchbox. As solving this issue might take further time and using the approach to do a 'export QT_GRAPHICSSYSTEM="raster"' might not be obvious, the graphicssystem is set manually to "raster" for the tests.