Changes: Isolate status bar animation code in a single class
RevBy: Marcin Miklas
MR: 2012

Details: And also removed hardcoded animation parameters along with it.

Previously if you wanted to change the duration or the easing cuve of
the status bar animation you would also have to do the same for
the MWidgetMoveAnimations used for the root elements and for the hardcoded
geometry animations used for pages and sheets. Now you will do it only
once in MStatusBarAnimationStyle.