Fixes: NB#269029 - MTF causes uninitialised reads
RevBy: Holger Schröder
MR: 2027

Details:
Valgrind was outputting some uninitialized values errors related
to a timer so this initializes it and fixes this warning.

QElapsedTimer is a POD, so it does not have a constructor that
gets called when it would be for e.g. a QString. so we have
to explicitly call the default constructor.