Fixes: NB#262511 - Regression - Dialog appears under ApplicationMenu
RevBy: Daniel d'Andrada
Details: Changed root elemenent for ApplicationMenu back to topNavigationBarRootElement.
zValue of topNavigationBarRootElement set to zValue of NavigationBar + 1, this is to prevent
ApplicationMenu to be displayed below NavigationBar.
It is because NavigationBar is added later to rootElement than topNavigationBarRootElement
and in case of same zValues later added items are displayed on top.