Changes: MApplicationWindow resets chained property when chained data is not available.
RevBy: Daniel d'Andrada
Details: If MComponentData::chainDataStackIsEmpty MApplicationWindow resets
_MEEGOTOUCH_WM_INVOKED_BY and TransientForHint properties.

This fixes case when application window was previously shown by dbus call containing
chain task data and now it is called differently and that chain task data is not available.
In this case we must resets chaining related x properties to prevent application window
displaying as part of application which called it last time.