New: MSheet::centralWidgetSizePolicyRespected property
Fixes: NB#272817

RevBy: Marcin Miklas
Details:
Currently central widget is always stretched out to fill the entire central area
of the sheet.
This is a problem when you have a sheet with litle content including one or more text edits,
having the bottom half of the central area free of widgets.
In this case, when a text edit gets focus and the virtual keyboard slides up a position indicator
will show up because you are able to pan downwards to access all the empty area covered by the
virtual keyboard. If the central widget in this example didn't get stretched to fill out the
central area that needless panning issue wouldn't happen.
This is now possible to achieve by setting MSheet::centralWidgetSizePolicyRespected property
to true.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2069