Fixes: NB#271740 - black content while a systemsheet is sliding away
RevBy: TrustMe
Details: mcompositor needs to get a _NET_CLOSE_WINDOW in order to be able
to get the contents of the window before animating it.

So when you make a systemwide sheet disappear() the MWindow that backs it
will be closed instead of hidden.

Fix based on idea by Michał Palczewski and Marcin Miklas.