Changes: Use qweakpointer in mbasiclistitemdeletionanimation to avoid operating on deleted objects.
RevBy: Daniel d'Andrada
MergeRequest: 2058

Details: mlistviewprivate deletes the animated widgets on animation finished signal, the animation itself was resetting target widget properties on the same finished signal.
Thus it was a race condition between deletion of the object by mlistviewprivate and animation which was trying to reset the targets widget state.
This commit ensures via qweakpointer that the deleted target widget object shall not be operated by the animation.