Fixes: NB#258946 - Clock; The name of city displays incorrectly in clock main view.
RevBy: Berthold Krevert

Details:

Put Capitals first in the return value of MLocationDatabase:citiesInTimeZone(),
before cities which match the Olson ID of the timezone.

Up to two cities are displayed in the timezone selection dialog.
Before this patch the first one was the city matching the Olson ID
and the second one was the capital (unless these are already identical).

After this patch, the order is just the other way round, the capital
comes first, then the city matching the Olson ID.

In Clock, only one city is displayed, before this patch it was the
city matching the Olson ID, now it is the capital.