Changes: Adding a forth way of specifying the animations of a scene window
RevBy: TrustMe
Details: Add the possibility for the creator of the scene window to specify
the animations to be used by it via internal dynamic properties.

It will have a hight priority than any animation definition in the scene window
style or scene manager style.

That way animations can be changed programmatically to better fit the need of the
moment, adding more flexibility.

Note that this will be used only internally and therefore is not part of the public API.