Fixes: NB#265473, Text magnifier is not working in Nokia account creation
RevBy: Daniel d'Andrada
Details: Custom focus handling in MScene was working incorrectly when
using embedded QML. The declarative items manage their own focus setting
via 'on pressed' handlers and QGraphicsScene stays out of it.

Graphics items have a special variable 'mouseSetsFocus' which tells whether
or not the item is focused with mouse. The variable is very similar to
QGraphicsItem flag ItemStopsFocusHandling in that whenever such an item is
encountered, QGraphicsScene's focus handling silently stops. With
QDeclarativeItems this variable is false, while by default it is true.

Downside is that mouseSetsFocus is not accessible outside QGraphicsItem,
and for now the value of the variable is assumed based on whether the item
inherits QDeclarativeItem.

https://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/2103